[Silver V] Title: 최소공배수, Time: 108 ms, Memory: 14252 KB -BaekjoonHub

+++ 백준/Silver/13241. 최소공배수/최소공배수.java
@@ -7,8 +7,8 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        long a = Integer.parseInt(st.nextToken());
-        long b = Integer.parseInt(st.nextToken());
+        long a = Long.parseLong(st.nextToken());
+        long b = Long.parseLong(st.nextToken());
 
         long lcm = a * b / gcd(a, b);
         System.out.print(lcm);
